+++ cpu-sources/sources/utility/image.cpp
@@ -16,11 +16,11 @@ shared_image load_png(const std::string filename)
 
     // Creating shared image
     shared_image sharedImage =
-        std::make_shared<Image<Pixel>>(size_t(image.cols), size_t(image.rows));
+        std::make_shared<Image<Pixel>>(int(image.cols), int(image.rows));
 
     // Copying data from cv::Mat to shared image
-    for (size_t y = 0; y < sharedImage->get_height(); y++)
-        for (size_t x = 0; x < sharedImage->get_width(); x++)
+    for (int y = 0; y < sharedImage->get_height(); y++)
+        for (int x = 0; x < sharedImage->get_width(); x++)
             sharedImage->set(x, y, rgbImage.at<Pixel>(y, x));
 
     return sharedImage;
@@ -30,8 +30,8 @@ void save_png(const std::string filename, shared_image image)
 {
     // Creating cv::Mat from shared image
     cv::Mat rgbImage(image->get_height(), image->get_width(), CV_8UC3);
-    for (size_t y = 0; y < image->get_height(); y++)
-        for (size_t x = 0; x < image->get_width(); x++)
+    for (int y = 0; y < image->get_height(); y++)
+        for (int x = 0; x < image->get_width(); x++)
             rgbImage.at<Pixel>(y, x) = image->get(x, y);
 
     // Saving image in BGR format
